# api_docs/openapi.yaml
openapi: 3.1.0
info:
  title: ComfyUI Orchestration Platform API
  version: "0.1.0"
  description: >
    Client-facing OpenAPI spec for the backend HTTP API. Base path is /api/v1.
servers:
  - url: http://localhost:8000
    description: Local dev
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    GenerationMode:
      type: string
      enum: [t2i, i2i, t2v, i2v]
    JobStatus:
      type: string
      enum: [queued, running, retrying, completed, failed, cancelled]
    LoraInjection:
      type: object
      required: [name]
      properties:
        name: { type: string, description: "LoRA filename available to ComfyUI." }
        strength_model: { type: number, default: 1.0, description: "Model strength for LoRA injection." }
        strength_clip: { type: number, default: 1.0, description: "CLIP strength for LoRA injection." }
    WorkflowRuntimeParameters:
      type: object
      properties:
        prompt:
          type: string
          default: a cinematic robot chef in a neon kitchen, high detail, soft studio lighting
          description: "Positive text prompt injected into Flux Schnell node 6."
        negative_prompt:
          type: [string, "null"]
          default: ""
          description: "Negative prompt injected into node 33. Flux Schnell commonly uses an empty negative prompt."
        seed:
          type: [integer, "null"]
          default: 173805153958730
          description: "KSampler seed for reproducible output. Change it for a different image."
        model:
          type: [string, "null"]
          default: flux1-schnell-fp8.safetensors
          description: "Checkpoint name injected into CheckpointLoaderSimple node 30."
        sampler: { type: [string, "null"], default: euler, description: "KSampler sampler_name for node 31." }
        scheduler: { type: [string, "null"], default: simple, description: "KSampler scheduler for node 31." }
        steps: { type: [integer, "null"], minimum: 1, default: 4, description: "Flux Schnell is designed for low step counts; default is 4." }
        cfg: { type: [number, "null"], minimum: 0, default: 1, description: "Classifier-free guidance scale. Flux Schnell default is 1." }
        denoise: { type: [number, "null"], minimum: 0, maximum: 1, default: 1, description: "Denoise strength for the KSampler." }
        width: { type: [integer, "null"], minimum: 1, default: 1024, description: "Generated image width injected into EmptySD3LatentImage node 27." }
        height: { type: [integer, "null"], minimum: 1, default: 1024, description: "Generated image height injected into EmptySD3LatentImage node 27." }
        batch_size: { type: [integer, "null"], minimum: 1, default: 1, description: "Number of images generated in one workflow run." }
        input_image: { type: [string, "null"], description: "Reserved for image-to-image workflows; unused by flux_schnell." }
        input_video: { type: [string, "null"], description: "Reserved for video workflows; unused by flux_schnell." }
        loras:
          type: array
          items: { $ref: "#/components/schemas/LoraInjection" }
          default: []
          description: "LoRA injections. Leave empty for flux_schnell because the workflow has no LoraLoader nodes."
        extra_parameters:
          type: object
          additionalProperties: true
          default: {}
          description: "Advanced node override map using node_id.input keys, for example {'31.steps': 6}."
    WorkflowExecutionRequest:
      type: object
      properties:
        mode:
          allOf:
            - $ref: "#/components/schemas/GenerationMode"
          default: t2i
          description: "Generation mode. Defaults to text-to-image."
        parameters:
          allOf:
            - $ref: "#/components/schemas/WorkflowRuntimeParameters"
          default: {}
          description: "Runtime parameters. Omit the object or individual fields to use Flux Schnell defaults."
        workflow_name:
          type: [string, "null"]
          default: flux_schnell
          description: "Workflow template filename without .json. Defaults to workflow/flux_schnell.json."
    OutputAsset:
      type: object
      required: [filename, subfolder, type, url, media_type]
      properties:
        filename: { type: string }
        subfolder: { type: string }
        type: { type: string }
        url: { type: string, description: "Absolute downloadable URL for HTTP API responses." }
        media_type: { type: string }
    ProgressEvent:
      type: object
      required: [job_id, status, created_at]
      properties:
        job_id: { type: string }
        prompt_id: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/JobStatus" }
        node_id: { type: [string, "null"] }
        current: { type: [integer, "null"] }
        maximum: { type: [integer, "null"] }
        message: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
    WorkflowExecutionResult:
      type: object
      required: [job_id, prompt_id, status, outputs, events, started_at]
      properties:
        job_id: { type: string }
        prompt_id: { type: string }
        status: { $ref: "#/components/schemas/JobStatus" }
        outputs:
          type: array
          items: { $ref: "#/components/schemas/OutputAsset" }
        events:
          type: array
          items: { $ref: "#/components/schemas/ProgressEvent" }
        started_at: { type: string, format: date-time }
        completed_at: { type: [string, "null"], format: date-time }
        error: { type: [string, "null"] }
    TrackedJob:
      type: object
      required: [job_id, mode, status, retry_count, outputs, created_at, updated_at]
      properties:
        job_id: { type: string }
        task_id: { type: [string, "null"] }
        prompt_id: { type: [string, "null"] }
        mode: { $ref: "#/components/schemas/GenerationMode" }
        status: { $ref: "#/components/schemas/JobStatus" }
        retry_count: { type: integer, default: 0 }
        progress_current: { type: [integer, "null"] }
        progress_maximum: { type: [integer, "null"] }
        error: { type: [string, "null"] }
        outputs:
          type: array
          items: { $ref: "#/components/schemas/OutputAsset" }
          default: []
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    QueueMonitoringResponse:
      type: object
      required:
        - queue_name
        - pending_count
        - active_count
        - reserved_count
        - scheduled_count
        - worker_concurrency
        - gpu_safe
        - persisted_status_counts
      properties:
        queue_name: { type: string }
        pending_count: { type: integer }
        active_count: { type: integer }
        reserved_count: { type: integer }
        scheduled_count: { type: integer }
        worker_concurrency: { type: integer }
        gpu_safe: { type: boolean }
        persisted_status_counts:
          type: object
          additionalProperties: { type: integer }
...
    JobListResponse:
      type: object
      required: [total, limit, offset, jobs]
      properties:
        total: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1 }
        offset: { type: integer, minimum: 0 }
        status: { anyOf: [{ $ref: "#/components/schemas/JobStatus" }, { type: "null" }] }
        jobs:
          type: array
          items: { $ref: "#/components/schemas/TrackedJob" }
    JobCountsResponse:
      type: object
      required: [counts]
      properties:
        counts:
          type: object
          additionalProperties: { type: integer }
    LogTailResponse:
      type: object
      required: [lines, truncated]
      properties:
        lines:
          type: array
          items: { type: string }
        truncated: { type: boolean }
  responses:
    UnauthorizedError:
      description: Missing/invalid credentials.
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/v1/generation/execute:
    post:
      tags: [generation]
      summary: Execute workflow immediately and await results.
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkflowExecutionRequest" }
            examples:
              fluxDefaults:
                summary: Use Flux Schnell defaults
                value: {}
              overridePrompt:
                summary: Override only selected defaults
                value:
                  parameters:
                    prompt: a cinematic robot chef in a neon kitchen, high detail, soft studio lighting
                    seed: 12345
              explicitDefaults:
                summary: Full Flux Schnell default request
                value:
                  mode: t2i
                  workflow_name: flux_schnell
                  parameters:
                    prompt: a cinematic robot chef in a neon kitchen, high detail, soft studio lighting
                    negative_prompt: ""
                    seed: 173805153958730
                    model: flux1-schnell-fp8.safetensors
                    sampler: euler
                    scheduler: simple
                    steps: 4
                    cfg: 1
                    denoise: 1
                    width: 1024
                    height: 1024
                    batch_size: 1
                    loras: []
                    extra_parameters: {}
      responses:
        "200":
          description: Execution result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkflowExecutionResult" }
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "502":
          description: Upstream ComfyUI error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail: { type: string }
...
  /api/v1/generation/jobs/{job_id}:
    get:
      tags: [generation]
      summary: Get tracked job by id.
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - in: path
          name: job_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Job.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrackedJob" }
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: Job not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail: { type: string }
  /api/v1/generation/jobs/{job_id}/cancel:
    post:
      tags: [generation]
      summary: Cancel a queued/running job.
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - in: path
          name: job_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated job.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrackedJob" }
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: Job not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail: { type: string }
        "502":
          description: Upstream ComfyUI error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail: { type: string }
  /api/v1/generation/queue:
    get:
      tags: [generation]
      summary: Monitor current queue snapshot.
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        "200":
          description: Queue snapshot.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QueueMonitoringResponse" }
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /api/v1/monitoring/jobs:
    get:
      tags: [monitoring]
      summary: List tracked jobs for dashboards.
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: status
          schema:
            anyOf:
              - $ref: "#/components/schemas/JobStatus"
              - type: "null"
      responses:
        "200":
          description: Job list.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobListResponse" }
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v1/monitoring/queue:
    get:
      tags: [monitoring]
      summary: Queue snapshot (same as generation queue monitor).
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        "200":
          description: Queue snapshot.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QueueMonitoringResponse" }
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v1/monitoring/jobs/counts:
    get:
      tags: [monitoring]
      summary: Counts of jobs by status.
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        "200":
          description: Counts map.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobCountsResponse" }
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v1/monitoring/logs/tail:
    get:
      tags: [monitoring]
      summary: Tail application log (admin only).
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - in: query
          name: lines
          schema: { type: integer, minimum: 1, maximum: 2000, default: 200 }
      responses:
        "200":
          description: Log tail lines.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LogTailResponse" }
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Admin privileges required.
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail: { type: string }